backend/server/requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0
psycopg2-binary==2.9.3
psycopg2==2.9.3
requests==2.26.0